move everglades to its own top-level package
