--- src/main/java/gtPlusPlus/everglades/biome/GenLayerBiomes.java
@@ -1,26 +1,0 @@
-package gtPlusPlus.everglades.biome;
-
-import net.minecraft.world.biome.BiomeGenBase;
-import net.minecraft.world.gen.layer.GenLayer;
-import net.minecraft.world.gen.layer.IntCache;
-
-public class GenLayerBiomes extends GenLayer {
-
-    protected BiomeGenBase[] allowedBiomes = { BiomeEverglades.biome, };
-
-    public GenLayerBiomes(long seed) {
-        super(seed);
-    }
-
-    @Override
-    public int[] getInts(int x, int z, int width, int depth) {
-        int[] dest = IntCache.getIntCache(width * depth);
-        for (int dz = 0; dz < depth; dz++) {
-            for (int dx = 0; dx < width; dx++) {
-                this.initChunkSeed(dx + x, dz + z);
-                dest[(dx + dz * width)] = this.allowedBiomes[nextInt(this.allowedBiomes.length)].biomeID;
-            }
-        }
-        return dest;
-    }
-}
