--- src/main/java/gtPlusPlus/everglades/block/BlockDarkWorldPollutedDirt.java
@@ -1,68 +1,0 @@
-package gtPlusPlus.everglades.block;
-
-import static gregtech.api.enums.Mods.Minecraft;
-
-import net.minecraft.block.BlockDirt;
-import net.minecraft.world.ColorizerGrass;
-import net.minecraft.world.IBlockAccess;
-
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
-import gtPlusPlus.api.interfaces.ITileTooltip;
-import gtPlusPlus.core.creative.AddToCreativeTab;
-
-public class BlockDarkWorldPollutedDirt extends BlockDirt implements ITileTooltip {
-
-    public BlockDarkWorldPollutedDirt() {
-        this.setCreativeTab(AddToCreativeTab.tabBOP);
-        this.setBlockName("blockDarkWorldGround2");
-        this.setHardness(0.5F);
-        this.setBlockTextureName(Minecraft.ID + ":" + "dirt");
-    }
-
-    @Override
-    @SideOnly(Side.CLIENT)
-    public int getBlockColor() {
-        double d0 = 0.5D;
-        double d1 = 1.0D;
-        return ColorizerGrass.getGrassColor(d0, d1);
-    }
-
-    /**
-     * Returns the color this block should be rendered. Used by leaves.
-     */
-    @Override
-    @SideOnly(Side.CLIENT)
-    public int getRenderColor(int p_149741_1_) {
-        return this.getBlockColor();
-    }
-
-    /**
-     * Returns a integer with hex for 0xrrggbb with this color multiplied against the blocks color. Note only called
-     * when first determining what to render.
-     */
-    @Override
-    @SideOnly(Side.CLIENT)
-    public int colorMultiplier(IBlockAccess p_149720_1_, int p_149720_2_, int p_149720_3_, int p_149720_4_) {
-        int l = 0;
-        int i1 = 0;
-        int j1 = 0;
-
-        for (int k1 = -1; k1 <= 1; ++k1) {
-            for (int l1 = -1; l1 <= 1; ++l1) {
-                int i2 = p_149720_1_.getBiomeGenForCoords(p_149720_2_ + l1, p_149720_4_ + k1)
-                    .getBiomeGrassColor(p_149720_2_ + l1, p_149720_3_, p_149720_4_ + k1);
-                l += (i2 & 16711680) >> 16;
-                i1 += (i2 & 65280) >> 8;
-                j1 += i2 & 255;
-            }
-        }
-
-        return (l / 9 & 255) << 16 | (i1 / 9 & 255) << 8 | j1 / 9 & 255;
-    }
-
-    @Override
-    public int getTooltipID() {
-        return 3;
-    }
-}
