+++ src/main/java/gtPlusPlus/xmod/gregtech/HandlerGT.java
@@ -7,7 +7,6 @@ import cpw.mods.fml.common.event.FMLLoadCompleteEvent;
 import gregtech.api.items.MetaGeneratedTool;
 import gregtech.api.util.GTConfig;
 import gtPlusPlus.core.handler.CompatHandler;
-import gtPlusPlus.everglades.gen.gt.WorldGen_GT;
 import gtPlusPlus.recipes.CokeAndPyrolyseOven;
 import gtPlusPlus.xmod.gregtech.api.enums.GregtechOrePrefixes.GT_Materials;
 import gtPlusPlus.xmod.gregtech.common.MetaGTProxy;
@@ -21,6 +20,7 @@ import gtPlusPlus.xmod.gregtech.loaders.recipe.RecipeLoaderMolecularTransformer;
 import gtPlusPlus.xmod.gregtech.loaders.recipe.RecipeLoaderTreeFarm;
 import gtPlusPlus.xmod.gregtech.registration.gregtech.GregtechAdvancedBoilers;
 import gtPlusPlus.xmod.gregtech.registration.gregtech.GregtechConduits;
+import toxiceverglades.gen.gt.WorldGen_GT;
 
 public class HandlerGT {
